refactor(connect): move tron method definitions to corresponding folder
