+++ packages/connect-common/src/types/api/tron/tronComposeTransaction.ts
@@ -1,0 +1,6 @@
+import type { TronComposeTransaction, TronComposedTransaction } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function tronComposeTransaction(
+    params: Params<TronComposeTransaction>,
+): Response<TronComposedTransaction>;
