+++ packages/connect-common/src/types/api/tron/tronGetAddress.ts
@@ -1,4 +1,4 @@
-import type { Address, BundledParams, GetAddress, Params, Response } from '../params';
+import type { Address, BundledParams, GetAddress, Params, Response } from '../../params';
 
 export declare function tronGetAddress(params: Params<GetAddress>): Response<Address>;
 export declare function tronGetAddress(params: BundledParams<GetAddress>): Response<Address[]>;
