+++ packages/connect-common/src/types/api/tron/tronSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { TronSignTransaction, TronSignedTx } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function tronSignTransaction(
+    params: Params<TronSignTransaction>,
+): Response<TronSignedTx>;
