--- packages/connect-common/src/types/api/tronComposeTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { TronComposeTransaction, TronComposedTransaction } from './tron';
-
-export declare function tronComposeTransaction(
-    params: Params<TronComposeTransaction>,
-): Response<TronComposedTransaction>;
