--- packages/connect-common/src/types/api/tronSignTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { TronSignTransaction, TronSignedTx } from './tron';
-
-export declare function tronSignTransaction(
-    params: Params<TronSignTransaction>,
-): Response<TronSignedTx>;
