+++ packages/connect-common/src/types/index.ts
@@ -17,7 +17,7 @@ export * from './api/monero/common';
 export * from './api/solana/common';
 export * from './api/stellar/common';
 export * from './api/tezos/common';
-export * from './api/tron';
+export * from './api/tron/common';
 export * from './api/nostr';
 
 // types used in @trezor/suite. if you need a type, reexport it from ./api/<method>
